requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiosqlite==0.19.0
python-dateutil==2.8.2
pdfplumber==0.10.3
beautifulsoup4==4.12.3
lxml==5.1.0
openpyxl==3.1.2
pandas==2.2.0
aiohttp==3.9.1
Flask
gunicorn
requests
